Updated updated test cases

+++ src/main/java/com/hcl/hackathon/repository/OrderRepository.java
@@ -2,9 +2,6 @@
 package com.hcl.hackathon.repository;
 
 import com.hcl.hackathon.entity.OrderInfo;
-import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
-import org.springframework.data.repository.PagingAndSortingRepository;
-import org.springframework.stereotype.Repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
@@ -14,7 +11,7 @@ import java.util.List;
 
 
 @Repository
-public interface OrderRepository extends JpaRepository<OrderInfo, Long> {
+public interface OrderRepository extends JpaRepository<OrderInfo, Integer> {
 
 
     List<OrderInfo> findByOrderStatus(String orderStatus);
